+++ cardano-crypto-tests/src/Test/Crypto/VRF.hs
@@ -25,102 +25,95 @@ import qualified Data.ByteString as BS
 import qualified Data.Char as Char
 import Data.Proxy (Proxy (..))
 import Data.Word (Word64, Word8)
+import GHC.Stack (HasCallStack)
 import qualified Text.ParserCombinators.ReadP as Parse
 import qualified Text.Read as Read
 
 import Paths_cardano_crypto_tests (getDataFileName)
 import Test.Crypto.Util
+import Test.HUnit (assertBool, assertFailure, (@?=))
+import Test.Hspec (Expectation, Spec, describe, it)
+import Test.Hspec.QuickCheck (prop)
 import Test.QuickCheck (
   Arbitrary (..),
   Gen,
   NonNegative (..),
   Property,
   counterexample,
+  vectorOf,
   (===),
   (==>),
  )
-import Test.Tasty (TestTree, testGroup)
-import Test.Tasty.HUnit (Assertion, HasCallStack, assertBool, assertFailure, testCase, (@?=))
-import Test.Tasty.QuickCheck (testProperty, vectorOf)
 
 {- HLINT IGNORE "Use <$>" -}
 --
 -- The list of all tests
 --
-tests :: TestTree
+tests :: Spec
 tests =
-  testGroup
-    "Crypto.VRF"
-    [ testVRFAlgorithm (Proxy :: Proxy MockVRF) "MockVRF"
-    , testVRFAlgorithm (Proxy :: Proxy SimpleVRF) "SimpleVRF"
-    , testVRFAlgorithm (Proxy :: Proxy PraosVRF) "PraosVRF"
-    , testVRFAlgorithm (Proxy :: Proxy PraosBatchCompatVRF) "PraosBatchCompatVRF"
-    , testGroup
-        "OutputVRF"
-        [ testProperty "bytesToNatural" prop_bytesToNatural
-        , testProperty "naturalToBytes" prop_naturalToBytes
-        ]
-    , testGroup
-        "ConvertingTypes"
-        [ testProperty "pubKeyToBatchCompat" prop_pubKeyToBatchComopat
-        , testProperty "signKeyToBatchCompat" prop_signKeyToBatchCompat
-        , testProperty "outputToBatchCompat" prop_outputToBatchComat
-        , testProperty "compatibleVerKeyConversion" prop_verKeyValidConversion
-        , testProperty "compatibleSignKeyConversion" prop_signKeyValidConversion
-        ]
-    , testGroup
-        "test vectors for Praos"
-        [ testCase "generated golden test vector: vrf_ver03_generated_1" $
-            checkVer03TestVector "vrf_ver03_generated_1"
-        , testCase "generated golden test vector: vrf_ver03_generated_2" $
-            checkVer03TestVector "vrf_ver03_generated_2"
-        , testCase "generated golden test vector: vrf_ver03_generated_3" $
-            checkVer03TestVector "vrf_ver03_generated_3"
-        , testCase "generated golden test vector: vrf_ver03_generated_4" $
-            checkVer03TestVector "vrf_ver03_generated_4"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
-          testCase "generated golden test vector: vrf_ver03_standard_10" $
-            checkVer03TestVector "vrf_ver03_standard_10"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
-          testCase "generated golden test vector: vrf_ver03_standard_11" $
-            checkVer03TestVector "vrf_ver03_standard_11"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
-          testCase "generated golden test vector: vrf_ver03_standard_12" $
-            checkVer03TestVector "vrf_ver03_standard_12"
-        ]
-    , testGroup
-        "test vectors for PraosBatchCompat"
-        [ testCase "generated golden test vector: vrf_ver13_generated_1" $
-            checkVer13TestVector "vrf_ver13_generated_1"
-        , testCase "generated golden test vector: vrf_ver13_generated_2" $
-            checkVer13TestVector "vrf_ver13_generated_2"
-        , testCase "generated golden test vector: vrf_ver13_generated_3" $
-            checkVer13TestVector "vrf_ver13_generated_3"
-        , testCase "generated golden test vector: vrf_ver13_generated_4" $
-            checkVer13TestVector "vrf_ver13_generated_4"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 10
-          -- pi = 7d9c633ffeee27349264cf5c667579fc583b4bda63ab71d001f89c10003ab46f14adf9a3cd8b8412d9038531e865c341cafa73589b023d14311c331a9ad15ff2fb37831e00f0acaa6d73bc9997b06501
-          testCase "generated golden test vector: vrf_ver13_standard_10" $
-            checkVer13TestVector "vrf_ver13_standard_10"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 11
-          -- pi = 47b327393ff2dd81336f8a2ef10339112401253b3c714eeda879f12c509072ef055b48372bb82efbdce8e10c8cb9a2f9d60e93908f93df1623ad78a86a028d6bc064dbfc75a6a57379ef855dc6733801
-          testCase "generated golden test vector: vrf_ver13_standard_11" $
-            checkVer13TestVector "vrf_ver13_standard_11"
-        , -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 12
-          -- pi = 926e895d308f5e328e7aa159c06eddbe56d06846abf5d98c2512235eaa57fdce35b46edfc655bc828d44ad09d1150f31374e7ef73027e14760d42e77341fe05467bb286cc2c9d7fde29120a0b2320d04
-          testCase "generated golden test vector: vrf_ver13_standard_12" $
-            checkVer13TestVector "vrf_ver13_standard_12"
-        ]
-    ]
+  describe "Crypto.VRF" $ do
+    testVRFAlgorithm (Proxy :: Proxy MockVRF) "MockVRF"
+    testVRFAlgorithm (Proxy :: Proxy SimpleVRF) "SimpleVRF"
+    testVRFAlgorithm (Proxy :: Proxy PraosVRF) "PraosVRF"
+    testVRFAlgorithm (Proxy :: Proxy PraosBatchCompatVRF) "PraosBatchCompatVRF"
+    describe "OutputVRF" $ do
+      prop "bytesToNatural" prop_bytesToNatural
+      prop "naturalToBytes" prop_naturalToBytes
+    describe "ConvertingTypes" $ do
+      prop "pubKeyToBatchCompat" prop_pubKeyToBatchComopat
+      prop "signKeyToBatchCompat" prop_signKeyToBatchCompat
+      prop "outputToBatchCompat" prop_outputToBatchComat
+      prop "compatibleVerKeyConversion" prop_verKeyValidConversion
+      prop "compatibleSignKeyConversion" prop_signKeyValidConversion
+    describe "test vectors for Praos" $ do
+      it "generated golden test vector: vrf_ver03_generated_1" $
+        checkVer03TestVector "vrf_ver03_generated_1"
+      it "generated golden test vector: vrf_ver03_generated_2" $
+        checkVer03TestVector "vrf_ver03_generated_2"
+      it "generated golden test vector: vrf_ver03_generated_3" $
+        checkVer03TestVector "vrf_ver03_generated_3"
+      it "generated golden test vector: vrf_ver03_generated_4" $
+        checkVer03TestVector "vrf_ver03_generated_4"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
+      it "generated golden test vector: vrf_ver03_standard_10" $
+        checkVer03TestVector "vrf_ver03_standard_10"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
+      it "generated golden test vector: vrf_ver03_standard_11" $
+        checkVer03TestVector "vrf_ver03_standard_11"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/03/ - Section A.4.
+      it "generated golden test vector: vrf_ver03_standard_12" $
+        checkVer03TestVector "vrf_ver03_standard_12"
 
-bytesEq :: HasCallStack => (a -> BS.ByteString) -> Maybe a -> a -> Assertion
+    describe "test vectors for PraosBatchCompat" $ do
+      it "generated golden test vector: vrf_ver13_generated_1" $
+        checkVer13TestVector "vrf_ver13_generated_1"
+      it "generated golden test vector: vrf_ver13_generated_2" $
+        checkVer13TestVector "vrf_ver13_generated_2"
+      it "generated golden test vector: vrf_ver13_generated_3" $
+        checkVer13TestVector "vrf_ver13_generated_3"
+      it "generated golden test vector: vrf_ver13_generated_4" $
+        checkVer13TestVector "vrf_ver13_generated_4"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 10
+      -- pi = 7d9c633ffeee27349264cf5c667579fc583b4bda63ab71d001f89c10003ab46f14adf9a3cd8b8412d9038531e865c341cafa73589b023d14311c331a9ad15ff2fb37831e00f0acaa6d73bc9997b06501
+      it "generated golden test vector: vrf_ver13_standard_10" $
+        checkVer13TestVector "vrf_ver13_standard_10"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 11
+      -- pi = 47b327393ff2dd81336f8a2ef10339112401253b3c714eeda879f12c509072ef055b48372bb82efbdce8e10c8cb9a2f9d60e93908f93df1623ad78a86a028d6bc064dbfc75a6a57379ef855dc6733801
+      it "generated golden test vector: vrf_ver13_standard_11" $
+        checkVer13TestVector "vrf_ver13_standard_11"
+      -- https://datatracker.ietf.org/doc/draft-irtf-cfrg-vrf/13/ - example 12
+      -- pi = 926e895d308f5e328e7aa159c06eddbe56d06846abf5d98c2512235eaa57fdce35b46edfc655bc828d44ad09d1150f31374e7ef73027e14760d42e77341fe05467bb286cc2c9d7fde29120a0b2320d04
+      it "generated golden test vector: vrf_ver13_standard_12" $
+        checkVer13TestVector "vrf_ver13_standard_12"
+
+bytesEq :: HasCallStack => (a -> BS.ByteString) -> Maybe a -> a -> Expectation
 bytesEq outputToBytes suppliedM expected = case suppliedM of
   Just supplied ->
     outputToBytes supplied @?= outputToBytes expected
   Nothing ->
     assertBool ("suppliedM in byteEq gave Nothing") False
 
-checkVer03TestVector :: FilePath -> Assertion
+checkVer03TestVector :: FilePath -> Expectation
 checkVer03TestVector file = do
   filename <- getDataFileName $ "test_vectors/" <> file
   str <- readFile filename
@@ -146,7 +139,7 @@ checkVer03TestVector file = do
   -- verify verKey proof msg -> hashed msg
   bytesEq Ver03.outputBytes (Ver03.verify verKey proof' testVectorMessage) hash'
 
-checkVer13TestVector :: FilePath -> Assertion
+checkVer13TestVector :: FilePath -> Expectation
 checkVer13TestVector file = do
   filename <- getDataFileName $ "test_vectors/" <> file
   str <- readFile filename
@@ -261,103 +254,89 @@ testVRFAlgorithm ::
   ) =>
   proxy v ->
   String ->
-  TestTree
+  Spec
 testVRFAlgorithm _ n =
-  testGroup
-    n
-    [ testGroup
-        "serialisation"
-        [ testGroup
-            "raw"
-            [ testProperty "VerKey" $
-                prop_raw_serialise @(VerKeyVRF v)
-                  rawSerialiseVerKeyVRF
-                  rawDeserialiseVerKeyVRF
-            , testProperty "SignKey" $
-                prop_raw_serialise @(SignKeyVRF v)
-                  rawSerialiseSignKeyVRF
-                  rawDeserialiseSignKeyVRF
-            , testProperty "Cert" $
-                prop_raw_serialise @(CertVRF v)
-                  rawSerialiseCertVRF
-                  rawDeserialiseCertVRF
-            ]
-        , testGroup
-            "size"
-            [ testProperty "VerKey" $
-                prop_size_serialise @(VerKeyVRF v)
-                  rawSerialiseVerKeyVRF
-                  (sizeVerKeyVRF (Proxy @v))
-            , testProperty "SignKey" $
-                prop_size_serialise @(SignKeyVRF v)
-                  rawSerialiseSignKeyVRF
-                  (sizeSignKeyVRF (Proxy @v))
-            , testProperty "Cert" $
-                prop_size_serialise @(CertVRF v)
-                  rawSerialiseCertVRF
-                  (sizeCertVRF (Proxy @v))
-            ]
-        , testGroup
-            "direct CBOR"
-            [ testProperty "VerKey" $
-                prop_cbor_with @(VerKeyVRF v)
-                  encodeVerKeyVRF
-                  decodeVerKeyVRF
-            , testProperty "SignKey" $
-                prop_cbor_with @(SignKeyVRF v)
-                  encodeSignKeyVRF
-                  decodeSignKeyVRF
-            , testProperty "Cert" $
-                prop_cbor_with @(CertVRF v)
-                  encodeCertVRF
-                  decodeCertVRF
-            ]
-        , testGroup
-            "To/FromCBOR class"
-            [ testProperty "VerKey" $ prop_cbor @(VerKeyVRF v)
-            , testProperty "SignKey" $ prop_cbor @(SignKeyVRF v)
-            , testProperty "Cert" $ prop_cbor @(CertVRF v)
-            ]
-        , testGroup
-            "ToCBOR size"
-            [ testProperty "VerKey" $ prop_cbor_size @(VerKeyVRF v)
-            , testProperty "SignKey" $ prop_cbor_size @(SignKeyVRF v)
-            , testProperty "Sig" $ prop_cbor_size @(CertVRF v)
-            ]
-        , testGroup
-            "direct matches class"
-            [ testProperty "VerKey" $
-                prop_cbor_direct_vs_class @(VerKeyVRF v)
-                  encodeVerKeyVRF
-            , testProperty "SignKey" $
-                prop_cbor_direct_vs_class @(SignKeyVRF v)
-                  encodeSignKeyVRF
-            , testProperty "Cert" $
-                prop_cbor_direct_vs_class @(CertVRF v)
-                  encodeCertVRF
-            ]
-        ]
-    , testGroup
-        "verify"
-        [ -- NOTE: we no longer test against maxVRF, because the maximum numeric
-          -- value isn't actually what we're interested in, as long as all
-          -- keys/hashes have the correct sizes, which 'prop_size_serialise'
-          -- tests already.
-          testProperty "verify positive" $ prop_vrf_verify_pos @v
-        , testProperty "verify negative" $ prop_vrf_verify_neg @v
-        ]
-    , testGroup
-        "output"
-        [ testProperty "sizeOutputVRF" $ prop_vrf_output_size @v
-        , testProperty "mkTestOutputVRF" $ prop_vrf_output_natural @v
-        ]
-    , testGroup
-        "NoThunks"
-        [ testProperty "VerKey" $ prop_no_thunks @(VerKeyVRF v)
-        , testProperty "SignKey" $ prop_no_thunks @(SignKeyVRF v)
-        , testProperty "Cert" $ prop_no_thunks @(CertVRF v)
-        ]
-    ]
+  describe n $ do
+    describe "serialisation" $ do
+      describe "raw" $ do
+        prop "VerKey" $
+          prop_raw_serialise @(VerKeyVRF v)
+            rawSerialiseVerKeyVRF
+            rawDeserialiseVerKeyVRF
+        prop "SignKey" $
+          prop_raw_serialise @(SignKeyVRF v)
+            rawSerialiseSignKeyVRF
+            rawDeserialiseSignKeyVRF
+        prop "Cert" $
+          prop_raw_serialise @(CertVRF v)
+            rawSerialiseCertVRF
+            rawDeserialiseCertVRF
+
+      describe "size" $ do
+        prop "VerKey" $
+          prop_size_serialise @(VerKeyVRF v)
+            rawSerialiseVerKeyVRF
+            (sizeVerKeyVRF (Proxy @v))
+        prop "SignKey" $
+          prop_size_serialise @(SignKeyVRF v)
+            rawSerialiseSignKeyVRF
+            (sizeSignKeyVRF (Proxy @v))
+        prop "Cert" $
+          prop_size_serialise @(CertVRF v)
+            rawSerialiseCertVRF
+            (sizeCertVRF (Proxy @v))
+
+      describe "direct CBOR" $ do
+        prop "VerKey" $
+          prop_cbor_with @(VerKeyVRF v)
+            encodeVerKeyVRF
+            decodeVerKeyVRF
+        prop "SignKey" $
+          prop_cbor_with @(SignKeyVRF v)
+            encodeSignKeyVRF
+            decodeSignKeyVRF
+        prop "Cert" $
+          prop_cbor_with @(CertVRF v)
+            encodeCertVRF
+            decodeCertVRF
+
+      describe "To/FromCBOR class" $ do
+        prop "VerKey" $ prop_cbor @(VerKeyVRF v)
+        prop "SignKey" $ prop_cbor @(SignKeyVRF v)
+        prop "Cert" $ prop_cbor @(CertVRF v)
+
+      describe "ToCBOR size" $ do
+        prop "VerKey" $ prop_cbor_size @(VerKeyVRF v)
+        prop "SignKey" $ prop_cbor_size @(SignKeyVRF v)
+        prop "Sig" $ prop_cbor_size @(CertVRF v)
+
+      describe "direct matches class" $ do
+        prop "VerKey" $
+          prop_cbor_direct_vs_class @(VerKeyVRF v)
+            encodeVerKeyVRF
+        prop "SignKey" $
+          prop_cbor_direct_vs_class @(SignKeyVRF v)
+            encodeSignKeyVRF
+        prop "Cert" $
+          prop_cbor_direct_vs_class @(CertVRF v)
+            encodeCertVRF
+
+    describe "verify" $ do
+      -- NOTE: we no longer test against maxVRF, because the maximum numeric
+      -- value isn't actually what we're interested in, as long as all
+      -- keys/hashes have the correct sizes, which 'prop_size_serialise'
+      -- tests already.
+      prop "verify positive" $ prop_vrf_verify_pos @v
+      prop "verify negative" $ prop_vrf_verify_neg @v
+
+    describe "output" $ do
+      prop "sizeOutputVRF" $ prop_vrf_output_size @v
+      prop "mkTestOutputVRF" $ prop_vrf_output_natural @v
+
+    describe "NoThunks" $ do
+      prop "VerKey" $ prop_no_thunks @(VerKeyVRF v)
+      prop "SignKey" $ prop_no_thunks @(SignKeyVRF v)
+      prop "Cert" $ prop_no_thunks @(CertVRF v)
 
 prop_vrf_verify_pos ::
   forall v.